+++ src/Methods/Misc.php
@@ -118,7 +118,7 @@ class Misc
             SmartString::\$dateTimeFormat        = 'Y-m-d H:i:s';   // Default dateTimeFormat() format
             SmartString::\$phoneFormat           = [                // Default phone number formats
                 ['digits' => 10, 'format' => '(###) ###-####'],
-                ['digits' => 11, 'format' => '1-###-###-####'],
+                ['digits' => 11, 'format' => '# (###) ###-####'],
             ];
 
         __TEXT__;